collector_datev/requirements.txt
# Existing CLI dependencies
playwright==1.46.0
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1
pytest==8.3.2
openpyxl==3.1.5
duckduckgo-search>=6.3.0
google-search-results==2.4.2
requests>=2.31.0
playwright-stealth>=1.0.6

# Flask Web Framework
flask>=3.0.0
flask-sqlalchemy>=3.1.0

# Database migrations (optional)
alembic>=1.13.0

# HTML Parsing (for BStBK scraper)
beautifulsoup4>=4.12.0

# AI Integration (OpenRouter)
cryptography>=42.0.0
httpx>=0.27.0
